Make the header value configurable
In the past, the only valid value for the X-Wikimedia-Debug header was '1',
which caused the request to be passed to mw1017.eqiad.wmnet. It is now possible
to specify the exact host to route the request to (from a set of pre-approved
hosts), so the extension needs to be updated to allow the user to specify a
custom value. To do this, we make the extension's browser action open a popup
on click, which has an input widget for the header value, as well as a toggle
button for turning header injection on or off.