background: Fix default value
Follows-up 02869ead89 which removed mw1017 in favour of mwdebug1001,
but did not update the default value. As such, in v1.7 of the Chrome extension,
the dropdown menu does not have any of the 4 available options selected by default.
Instead, the loading of the selection on startup leaves the dropdown menu blank.