Minor design update
Enforce that the button is to the right of the options,
like it was in the original design. The original design
was implemented by giving the options no explicit width,
and floating the 70px button to the right. This had the desired
effect so long as the options were a single line less than
190px wide but that stopped being the case when PHP7 was added.

Restore it by moving the button higher up in the HTML, and
explicitly limiting the options DIV width to 190px.