Add support for 'Find in XHGui' footer link in Beta Cluster
As of today, the Beta Cluster has a working XHGui server,
and the beta version of the perf site proxies it similar to how
it does in production.

I've also deployed a Puppet patch to Beta that installs Tideways
on its app servers, and enabled xhgui in wmf-config/LabsServices.

All works now!

Ref https://phabricator.wikimedia.org/T180761.