Use x-request-id header instead of mw.config wgRequestId
This is identical on page views, and has the benefit of working even
when debugging non-HTML requests. For example, an api.php or load.php
request which has the header, but can't have an HTML <script> tag
in the response containing wgRequestId.

Another benefit is we can remove wgRequestId from the page view HTML,
since afaik it was mainly introduced for WikimediaDebug, which no
longer needs it after this. For manual debugging, the headers can
be used, which per the above, are more generally useful.

Bug: T279211
Change-Id: I2676cb0e0d31412be4866dc9831e139f9ffb726f