popup: Add early response check in content_script.js callback
Follows-up 296a2491d54d2ff, which added an optimization to not send
a response object from background.js when WikimediaDebug is turned off.

This way, we minimise overhead (and profiling noise) when in the
common case of browsing Wikipedia with WikimediaDebug off.

The problem is that while it is optional for the background process
to respond, the browser generally fills in a default response of
undefined if none of the processes send a response. This error is
only observable in Chrome currently due to an upstream Firefox bug.

Bug: T294335
Change-Id: Ib3bfde1c3134953a080af366f95c7ec3d00aead9