popup: Move getRealm to background.js
I didn't realize before, but we can actually run call tabs.query()
from background.js as well, which has the benefit of not having to
transfer `urlPatterns` from background to popup.

And, as used in the next patch, allows us to re-use the disallow
list for the footer logic in content_script.js, which also varies
by realm.

Two other minor changes:

* Remove the `window.addEventListener('load')` delay. This was here
  because the UI is parsed after the script tag in popup.html.
  Instead, move the script tag. Moving the script tag does not relay
  the browser request (pre-fetched by browser lookahead), and makes
  it naturally run the script at the right time without additional
  delays.

* Document why we have the set-theme workaround, by pointing to
  the bugzilla task I just filed for the underlying issue.

Change-Id: I24a1d735d944feb6cfb0ea923bb8100e144ead5c