popup: Use visibility instead of display for hiding the popup
This has the benefit of letting the browser immediately render the
popup at its regular size, which takes maybe 100ms to appear and
animate. By that time, we typically have everything ready and can
unhide the body content.

Without this change, we force the popup to stay hidden for longer,
which if the loading is slow for some reason makes it seem like the
button didn't work.

With this change, the popup will appear immediately, and usually
filled in right away as well. If it's slow for some reason, it'll
appear but initially blank until it's ready, which seems like a
smoother UX.

Change-Id: I0d97d6128f015e12734fdd86026ed9f66bda92aa