popup: Let options continue to render under the main toggle
Previously, the popup was split into two columns. One for the select
dropdown and the all options, and one on the right for the main toggle
on the top right. This meant that the space under the toggle was not
available to render additional options.

With the added PHP 7.4 option (Ic2dd7ad3a41e, T312653) this needlessly
creates a third row.

Fix by giving the select and main toggle their own "row", and let the
options be on their own row underneath that.

Also while at it, switch to CSS flexbox and make the 10px gap between
the select and toggle explicit. Previously it was implicit by virtue
of the body being 260px, "column" being 190px, and "aside" being 60px,
which left 10px unacounted for between them. We now use flexbox with
10px gap. Also get rid of the "margin-top: 5px" hack which I used to
vertically align the select and toggle as they aren't the exact same
height. This is now taken care of by "align-items: center".

Bug: T312653
Change-Id: If65e9ef635f10cb1431e52c6bbd7dd44373ba483