popup: Use CSS grid to place two options per row
In the previous release, the layout was changed to accomodate
two additional options. That was done by hand that time,
and now that we've removed those two options again, change it back
to two per line instead of what would otherwise happen which is
three on one line and a lone one below that.

To make future changes easier, adopt CSS grid. Then in the future we
can simply change "1fr 1fr" to "1fr 1fr 1fr" accordingly to change
how many items per row should be displayed.

I considered using flexbox as well, but that doesn't appear to have
an easy way to force the before-last item to join a (lone) last
item on an otherwise empty row. If that's possible, then that would
remove the need for manual layout changes entirely.

Change-Id: I401807fddacaea8bd623997655679eba259e21ae