popup: Create new "output" area with last five xhgui/logstash links
== Why ==

We're soon going to add "Excimer UI" output links as per T291015
that link to a per-request generated flame graph. In order to make
these discoverable on non-HTML documents (e.g. profiling an API or
load.php response), we want to render these within the popup.

This change creates the system for that link to be a part of, and
converts the two existing output types (xhgui, logstash) to adopt
that system.

== Status quo ==

* content_script.js was injected into the tab's main document.
* This sent a "content-script" message to background.js,
  which responded with current form state (i.e. is XWD enabled,
  and are xhprof/log ticked), and the realm based on URL (beta vs prod),
  and what the observed reqId was when the tab loaded.
* The response was then used to insert additional links into the
  skin footer at the bottom of the page.

For it to know the reqId, though, it had to have kept track of this
from before the tab was loaded, and associate it with the tabId.

== Change ==

The minimal change would have been to set up a similar tab tracking
system and render those links in the popup in a fixed place. I
decided to do it a little different for three reasons:

1. Doing so would alter/break the mental model of the popup.
   Instead of being a window into "the" state of extension, it would
   vary by page.

2. When the link is in the popup, it will be much easier to lose track
   of these as we can have only 1 at a time, whereas previously one
   could mentally track a given tab in the browser and trust that you
   can go back to it anytime and still open the link in the footer.

3. This tab tracking felt messy and more complex than needed, and I
   realized a different way of doing it would solve the above and
   create a better user experience while at it.

Whenever background.js observes an HTTP response, it adds xhgui/log
links to an array, keeping upto 5 items.

Whenever popup.html is opened or when the array changes, it renders
the HTML list from that array.

Bug: T291015
Change-Id: I5ab566337784a7072349ca2755be84f2a944f227