popup: Fix unwanted animation during popup open in Chrome
In Chrome, the message event is processed a handful of animation
frames later than in Firefox. The result is that, when popup.js is
applying the form state as remembered by background.js, it is
triggering the CSS animation every single time, as if a human had
just toggled the elements. Steps to reproduce this:

* Open the popup in Chrome.
* Turn the big button on.
* Close and re-open the popup, observe bug.
* Repeat, observe bug again (etc).

Avoid this by 1) disabling the transition when body-hidden is active
and 2) delaying its removal until at least 1 frame after popup.js
applies its state.

The pattern of calling rAF() twice is common in frontend JS, and
the reason is as follows:

* During the event loop tick where popup.js applies the form state,
  we can't remove className synchronously in Chrome because it will
  process all DOM writes together once JS yields. The fact that we
  remove the class "after" changing the form state does not change
  the fact that it will render the changed form in the same paint
  frame where the body-hidden class is no longer applied and so
  an animation would happen.

* Calling rAF() once brings us from the main JS portion of a single
  event loop tick, to the end of the event loop tick right before
  the paint action takes place. Given nothing has happened in-between
  and given no third party JS here, this is effectively the same point
  in time still.

* Calling rAF() twice thus brings you to right after the form state
  has been rendered without a transition, and now we can safely remove
  the body-hidden class.

The artificial delay is about ~16ms (1/60fps) on modern screens and
is not expected to introduce annoyance in the form of clicks being
ignored.

Change-Id: If549d9120234a9a3838abc5826563eedb8efba19