popup: Fix darkmode
Follows-up I5ab566337784a70, which added the new output area, but
it hardcoded lightmode colors. In darkmode, it was a mess.

Avoid this in the future by not having inlined color and manually
kept colors in a `@media` block. Instead, de-duplicate colors into
a small set of variables, set those correclty for darkmode, and them
freely re-use those variables wherever we need them.

Change-Id: I1141fee439c8bc598bab86cb7d6cc74c52530442