Convert to Manifest v3
Convert from Manifest v2 to Manifest v3 for continued support of Chrome
which is attempting to remove Manifest v2 support.

The major difference in the v3 implementation is that rather than
directly inspecting and modifying outbound web requests we use the new
`declarativeNetRequest` rules system to register a rule explaining the
header we would like to add and the conditions necessary to trigger
adding it.

In Chrome the business logic from the former `background.js` script is
now executed as a service worker process. This change makes it a bit more
complicated to inspect and manipulate the browser's UI. For the moment
this means we have lost light/dark mode control of the toolbar icon in
Chrome. Firefox interestingly has a config based solution for this.

== Cross-browser differences ==
=== Background script ===
* Chrome doesn't support the `background.scripts` setting in v3. It
  instead requires a `background.service_worker` setting and conversion of
  background scripts to a service worker.
* Firefox doesn't support the `background.service_worker` setting in v3.
  It instead emits a "background.service_worker is currently disabled"
  fatal error.
* Setting both seems to work in the same horrible way that CSS things
  that are unsupported are often ignored after emitting a warning. See
  <https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/manifest.json/background#browser_support>

== Icons ==
* Firefox has support for `action.theme_icons` to configure light/dark
  mode, __BUT__ in super confusing news in that config "light" means "used
  where light text is needed" rather than "used where the background is
  light colored". Basically "light"/"dark" are the opposite of what you
  might reasonably expect.
* Chrome does not support `action.theme_icons`. They are proposing
  <https://github.com/w3c/webextensions/blob/main/proposals/dark_mode_extension_icons.md>
  upstream. See also <https://issues.chromium.org/issues/41419485>.

== Rule resource types ==
`Object.values( chrome.declarativeNetRequest.ResourceType )` returns 15
values in Chrome and 20 in Firefox. Neither list is a strict subset of
the other. The only reason this matters is that at least in Chrome (not
tested by me in Firefox) `declarativeNetRequest` rules do not seem to
apply to top level navigation (browser bar URLs, clicking links) unless
`condition.resourceTypes` includes "main_frame". This is not documented
anywhere that I can find in the official docs. As both browser runtimes
support querying for their supported types we do that at runtime rather
than hard coding a list that is subset or superset of the two
collections.

Bug: T312694
Change-Id: Id21263b3ab489a516d9ddf2a92d87cfacf30d1e6