Avoid block content on any page permission
As part of converting to manifest v3 we have added the permission for
`declarativeNetRequest`. It allows blocking or upgrading any requests
regardless of host permissions. On Chrome it triggers an installation
warning.

An alternative is `declarativeNetRequestWithHostAccess` which does not
trigger a warning which requires host permissions to be set which we
already do.

When inspecting the extension under chrome://extensions it states:

  Permissions
  * Read your browsing history
  * Block content on any page

After changing the permission, UNLOADING and reloading the extension:

  Permissions
  * Read your browsing history

Given the permission is gone, I expect the installation warning to
vanish.

I have verified on both Chromium and Firefox:
* the rules are injected (from the service worker console)
* when enabled, a request is responded with Server:
  mwdebug1001.eqiad.wmnet

Bug: T387822
Change-Id: I86d63c262e33e163831c7827a64f307477f502fe